requirements.txt
alembic==1.17.1
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
bcrypt==4.1.2
cffi==2.0.0
click==8.3.0
colorama==0.4.6
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
ecdsa==0.19.1
et-xmlfile==2.0.0
fastapi==0.120.4
fonttools==4.60.1
greenlet==3.2.4
h11==0.16.0
idna==3.11
jose==1.0.0
kiwisolver==1.4.9
lxml==6.0.2
mako==1.3.10
markupsafe==3.0.3
matplotlib==3.10.7
numpy==2.3.4
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
passlib==1.7.4
pillow==12.0.0
pyarrow==22.0.0
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.3
pydantic-core==2.41.4
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.20
pytz==2025.2
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
sqlalchemy==2.0.44
sqlmodel==0.0.27
starlette==0.49.1
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.2
uvicorn==0.38.0
